4.9.112.5
Added Creative ESPE Source
Added a fire effect to Orbital Strike
Added commands to set wind relations
Added wrong usage message to commands
Changed Captured Soul to use NBT
Changed all recipes to be Ingredient-based
Changed all loot to use loot tables
Nerfed Hoanna loot because MC-91727
Changed entity registry names
Renamed some fields
Changed Ultra Heat Generator calulations to be continuous
Fixed Research Book GUI
Fixed CraftTweaker integration
Fixed Town generation

+++ src/main/java/essentialcraft/client/render/entity/RenderMRUArrow.java
@@ -41,7 +41,7 @@ public class RenderMRUArrow extends Render<EntityMRUArrow> {
 		float f7 = 0.3125F;
 		float f8 = 0.05625F;
 		GlStateManager.enableRescaleNormal();
-		float f9 = (float)entity.arrowShake - partialTicks;
+		float f9 = entity.arrowShake - partialTicks;
 
 		if(f9 > 0.0F) {
 			float f10 = -MathHelper.sin(f9 * 3.0F) * f9;
